chore: renamed module from login to auth
