+++ package.json
@@ -32,9 +32,11 @@
     "prisma": "^4.5.0"
   },
   "dependencies": {
+    "@nestjs/jwt": "^9.0.0",
     "@prisma/client": "^4.5.0",
     "bcrypt": "^5.1.0",
     "crypto-js": "^4.1.1",
+    "passport-jwt": "^4.0.0",
     "randomstring": "^1.2.3"
   },
   "type": "module"
